+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # changelog
 
+## 2.2.0+27
+
+- start adding playback queue for songs.
+
 ## 2.1.0+26
 
-- add scroble util and integrate it to song play.
+- add scrobble util and integrate it to song play.
 
 ## 2.0.0+25